resolved review, refactor code and reused component
