+++ src/components/Button/Button.scss
@@ -1,0 +1,5 @@
+.custom-button {
+  border: none;
+  outline: none;
+  background-color: transparent;
+}
