+++ src/components/ImageDnD/ImageDnD.scss
@@ -1,0 +1,15 @@
+.container {
+  height: 10rem;
+  padding: 1rem;
+  border-radius: 2px;
+  border-color: #e8e8ea;
+  border-style: dashed;
+  background-color: #ffffff;
+  color: #bdbdbd;
+}
+
+.image-preview {
+  display: block;
+  width: auto;
+  max-height: 4rem;
+}
