--- src/components/ProtectedRoute/ProtectedRoute.js
@@ -1,8 +1,0 @@
-import React, { useState } from "react";
-
-const ProtectedRoute = () => {
-  const [isAutu, setIsAuth] = useState(false);
-  return <div>ProtectedRoute</div>;
-};
-
-export default ProtectedRoute;
