+++ src/main.jsx
@@ -6,9 +6,9 @@ import store from "./rtk/app/store.js";
 import { Provider } from "react-redux";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>
+  <Provider store={store}>
+    <App />
+  </Provider>
+  // <React.StrictMode>
+  // </React.StrictMode>
 );
